Blue-Submarine  gnome-applications.css to gtk3.20
Port gnome-applications.scc to gtk3.20, same file as Green-Submarine (colors are in gtk.css and settings.ini only). This work is preliminary, focussed only on suppressing errors. Much remains to do for GNOME applications due to the old selectors not being recognized, but GNOME might fix that themselves for their own apps