Blue-Submarine  other-applications.css to gtk3.20
Blue-Submarine  Port other-applications.css to gtk3.20 same file as Green-Submarine (colors are in gtk.css and settings.ini only). Focus is on suppressing errors and a rough match of Nemo to Caja. Note that the first three Nemo toolbar buttons on the left still jump on hover, need to explicitly set same border width all states.