fix cpufreq applet label problem
The cpufreq-applet units label was not coming up at panel startup, and a bottom panel was rendering 1px above the bottom of the screen. Fix both of these by overriding the theme default default 1px padding for everything in the panel. BlackMATE is one of the few themes with a non-zero default padding, it seems to cause bugs with applet and panel rendering not found in otther themes.