Fix HighContrast themes visibility with metacity
When rendered with metacity (e.g. with metacity-theme-viewer), the
back background parts of HighContrast and HighContrastInverse actually
show up transparent. This is because the corresponding rectangles were
missing the filled attribute.

In the HighContrast case, the gtk_arrow is getting drawn black on black (and
there is currently no way to change the color), so we can as well draw
it by hand to be able to change the color.

The close button also deserves bigger width to be more visible.

Fixes #211