HighContrast: Rework colors
We want the HightContrast theme to be as much "black-on-white" as
possible, so make buttons black on white by making the outerbevel filled
white (and only in the title area and borders, there is no need to fill
the whole window).  The currently-focused window keeps a white titlebar
while the other windows get a greyed titlebar.