Traditional and Menta theme families:  Resolved Caja sidebar display regressions
Several lines in these themes, seemingly related to historical overlay
scrolling issues, have been removed.  These lines were causing the doubling
in height of some rows in the Caja Places, Tree, History and Bookmarks
sidebars in all themes, and the chopping off of the left side of said
Caja sidebars in the Traditional family of themes.