Quid: Re-add legacy theme
This theme used to be called Crux, later renamed to Quid, and eventually
removed. This patch brings it back with some updated components. It does
not provide a full desktop theme (i.e. no colours or gtk widget
controls), but just the window decorations and icon theme.

Some distros provide a theme called Crux that only provides a gtkrc
file, and so GTK widgets will just default to Adwaita.